practice_const with double

+++ week_2/literal/main.c
@@ -4,7 +4,8 @@
 int main()
 {
     const unsigned long long con1 = 10;
-    const float con2 = 3.2f;
+    //const float con2 = 3.2f;
+    const double con2 = 3.2;
     const char con3 = 't';
     
     printf("%llu %f %c\n", con1, con2, con3);
